Consolidate DateMeter and DateTimeMeter into common source files
