--- DateMeter.h
@@ -1,15 +1,0 @@
-#ifndef HEADER_DateMeter
-#define HEADER_DateMeter
-/*
-htop - DateMeter.h
-(C) 2004-2011 Hisham H. Muhammad
-Released under the GNU GPLv2+, see the COPYING file
-in the source distribution for its full text.
-*/
-
-#include "Meter.h"
-
-
-extern const MeterClass DateMeter_class;
-
-#endif
