+++ DateTimeMeter.h
@@ -10,6 +10,8 @@ in the source distribution for its full text.
 #include "Meter.h"
 
 
+extern const MeterClass DateMeter_class;
+
 extern const MeterClass DateTimeMeter_class;
 
 #endif
