+++ linux/Platform.c
@@ -26,7 +26,6 @@ in the source distribution for its full text.
 #include "BatteryMeter.h"
 #include "ClockMeter.h"
 #include "CPUMeter.h"
-#include "DateMeter.h"
 #include "DateTimeMeter.h"
 #include "DiskIOMeter.h"
 #include "FileDescriptorMeter.h"
