+++ solaris/Platform.c
@@ -31,7 +31,6 @@ in the source distribution for its full text.
 #include "TasksMeter.h"
 #include "LoadAverageMeter.h"
 #include "ClockMeter.h"
-#include "DateMeter.h"
 #include "DateTimeMeter.h"
 #include "HostnameMeter.h"
 #include "SysArchMeter.h"
